core: split stdcm and static conflict detection
Not used for now and duplicates code, but it enables optimizations
